+++ pkg/checks/disk/space_test.go
@@ -1,16 +1,16 @@
-package checks_test
+package disk_test
 
 import (
 	"regexp"
 	"testing"
 
-	"github.com/conjurinc/conjur-preflight/pkg/checks"
+	"github.com/conjurinc/conjur-preflight/pkg/checks/disk"
 	"github.com/conjurinc/conjur-preflight/pkg/framework"
 	"github.com/stretchr/testify/assert"
 )
 
-func TestDiskRun(t *testing.T) {
-	testCheck := &checks.DiskSpace{}
+func TestSpaceCheck(t *testing.T) {
+	testCheck := &disk.SpaceCheck{}
 	resultChan := testCheck.Run()
 	results := <-resultChan