+++ pkg/cmd/root.go
@@ -2,20 +2,30 @@ package cmd
 
 import (
 	"fmt"
+	"io"
 	"os"
 
 	"github.com/conjurinc/conjur-preflight/pkg/framework"
+	"github.com/conjurinc/conjur-preflight/pkg/log"
 	"github.com/conjurinc/conjur-preflight/pkg/report"
 	"github.com/conjurinc/conjur-preflight/pkg/version"
 	"github.com/spf13/cobra"
 )
 
 func newRootCommand() *cobra.Command {
+	var debug bool
+
 	rootCmd := &cobra.Command{
 		Use:   "conjur-preflight",
 		Short: "Qualification CLI for common Conjur Enterprise self-hosted issues",
 		RunE: func(cmd *cobra.Command, args []string) error {
-			report := report.NewDefaultReport()
+			if debug {
+				log.EnableDebugMode()
+			}
+
+			report := report.NewDefaultReport(debug)
+
+			log.Debug("Running report...")
 			result := report.Run()
 
 			// Determine whether we want to use rich text or plain text based on
@@ -23,8 +33,10 @@ func newRootCommand() *cobra.Command {
 			o, _ := os.Stdout.Stat()
 			var formatStrategy framework.FormatStrategy
 			if (o.Mode() & os.ModeCharDevice) == os.ModeCharDevice { //Terminal
+				log.Debug("Using rich text report formatting")
 				formatStrategy = &framework.RichTextFormatStrategy{}
 			} else { //It is not the terminal
+				log.Debug("Using plain text report formatting")
 				formatStrategy = &framework.PlainTextFormatStrategy{}
 			}
 
@@ -33,12 +45,20 @@ func newRootCommand() *cobra.Command {
 				return err
 			}
 
-			fmt.Println(reportText)
-
+			fmt.Fprintln(cmd.OutOrStdout(), reportText)
+			log.Debug("Preflight finished!")
 			return nil
 		},
 		Version: version.FullVersionName,
 	}
+
+	rootCmd.PersistentFlags().BoolVarP(
+		&debug,
+		"debug",
+		"",
+		false,
+		"debug logging output",
+	)
 
 	// TODO: Add JSON output option
 	// TODO: Verbose logging control
@@ -49,13 +69,14 @@ func newRootCommand() *cobra.Command {
 
 // Execute adds all child commands to the root command and sets flags appropriately.
 // This is called by main.main(). It only needs to happen once to the rootCmd.
-func Execute() {
-	rootCmd.SetOut(os.Stdout)
-	rootCmd.SetErr(os.Stderr)
+func Execute(stdout, stderr io.Writer) {
+	rootCmd.SetOut(stdout)
+	rootCmd.SetErr(stderr)
 
 	err := rootCmd.Execute()
 
 	if err != nil {
+		log.Error("ERROR: %s\n", err)
 		os.Exit(1)
 	}
 }